+++ src/Login.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import './App.css';
+import { Container } from 'react-bootstrap';
+
+const Login = () => {
+    return (
+        <>
+            <Container id="login">
+                <h1>환영합니다!</h1>
+                <h1><span>로그인</span>을 부탁드려요!</h1>
+            </Container>
+        </>
+    );
+}
+
+export default Login;
