--- src/pages/CheckCallDetail.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-import { useParams } from "react-router-dom";
-
-export default function CheckCallDetail() {
-  const { id } = useParams();
-  return <div>{"CheckCallDetail, id: " + id}</div>;
-}
