--- src/pages/UploadCall.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function UploadCall() {
-  return <div>상담 업로드 페이지</div>;
-}
